Student Server --incomplete

+++ SpringBoot01/src/main/java/com/example/demo/StudentServer/Student.java
@@ -6,16 +6,16 @@ import jakarta.persistence.Id;
 @Entity
 public class Student {
     @Id
-    long id;
+    int id;
     String name;
     int age;
     String department;
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
